+++ src/main/java/com/newrelic/graphql/schema/scalars/DateTime.java
@@ -4,7 +4,10 @@
  */
 package com.newrelic.graphql.schema.scalars;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.util.Map;
 
 /**
  * ISO-8601 compliant custom scalar. Expects String inputs, converts to wrap a {@code ZonedDateTime}
@@ -16,6 +19,40 @@ public class DateTime {
   /** @param dateTime Construct with a specific datetime */
   public DateTime(ZonedDateTime dateTime) {
     this.dateTime = dateTime;
+  }
+
+  // Support for serialization from instances of our own type
+  @JsonCreator
+  private DateTime(Map<String, Object> props) {
+    this(rawValueFromProps(props));
+  }
+
+  protected static ZonedDateTime rawValueFromProps(Map<String, Object> props) {
+    Map<String, Object> value = cast(props.get("dateTime"));
+    if (value == null) {
+      throw new IllegalArgumentException("Can't deserialize from properties missing 'dateTime'");
+    }
+
+    Map<String, Object> zone = cast(value.get("zone"));
+    if (zone == null) {
+      throw new IllegalArgumentException(
+          "Can't deserialize from properties missing 'dateTime.zone'");
+    }
+
+    return ZonedDateTime.of(
+        (Integer) value.get("year"),
+        (Integer) value.get("monthValue"),
+        (Integer) value.get("dayOfMonth"),
+        (Integer) value.get("hour"),
+        (Integer) value.get("minute"),
+        (Integer) value.get("second"),
+        (Integer) value.get("nano"),
+        ZoneId.of((String) zone.get("id")));
+  }
+
+  @SuppressWarnings("unchecked")
+  private static <T> T cast(Object rawValue) {
+    return (T) rawValue;
   }
 
   /** @return Wrapped datetime object */